폴더 정리 및 yml 파일 변경, XML 매퍼에서 어노테이션 매퍼로 전환
