--- shoppingmall/src/main/resources/application.properties
@@ -1,20 +1,0 @@
-spring.application.name=shoppingmall
-server.port = 8080
-
-spring.datasource.url=ENC(jjh9WWdB0z22gcet0w3K1z5HQqqC4/bK9BL2TOo64b5IzUB/hgv1HPeUsl3Lv0JvHWd5SvtpsVvoLcCc0sNKPuckpJVyvDaXuLd1gFAwbAc=)
-spring.datasource.username=ENC(bdEJoVuvInRp4Ng+5FDT3qsy5HOMq6UQ244H+EGffRSKnhHzMScetHSHt+bI8V9Vnzfjes2r+zWY/NTKoCcBzw==)
-spring.datasource.password=ENC(96YolAAQfhHhjz04sS+qn1BaBnsV4mJLxhcKtMNcGqKbf8dw5DVLDN1ckXiUjGMr)
-spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
-mybatis.mapper-locations=classpath:mapper/*.xml
-
-#Thymeleaf
-spring.thymeleaf.cache=false
-spring.thymeleaf.enabled=true
-spring.thymeleaf.prefix=classpath:/templates/
-spring.thymeleaf.suffix=.html
-
-
-#Dubug
-#logging.level.org.jasypt=DEBUG
-#logging.level.org.springframework=DEBUG
-logging.level.com.ulisesbocchio.jasyptspringboot=DEBUG
