--- shoppingmall/src/main/resources/mapper/ProductMapper.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.lys.shoppingmall.mapper.ProductMapper">
-    <select id="getAllProducts" resultType="com.lys.shoppingmall.model.Product">
-        SELECT * FROM products;
-    </select>
-
-    <select id="getProductById" resultType="com.lys.shoppingmall.model.Product">
-        SELECT * FROM products WHERE id = #{id};
-    </select>
-
-    <insert id="insertProduct" parameterType="com.lys.shoppingmall.model.Product" useGeneratedKeys="true" keyProperty="id">
-        INSERT INTO products (name, price, stock) VALUES (#{name}, #{price}, #{stock});
-    </insert>
-
-    <update id="updateProduct" parameterType="com.lys.shoppingmall.model.Product">
-        UPDATE products
-        SET name = #{name}, price = #{price}, stock = ${stock}
-        WHERE id = #{id}
-    </update>
-
-    <delete id="deleteProduct">
-        DELETE FROM products WHERE id = #{id};
-    </delete>
-</mapper>
