--- shoppingmall/src/main/resources/static/hello.html
@@ -1,10 +1,0 @@
-<!-- resources/templates/hello.html -->
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-  <title>Hello</title>
-</head>
-<body>
-<h1>Hello, Thymeleaf!</h1>
-</body>
-</html>
