--- shoppingmall/src/main/resources/static/home.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-  <h2>
-    <a href="#">Click</a>
-  </h2>
-</body>
-</html>
