--- shoppingmall/src/main/resources/static/index.html
@@ -1,56 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Index</title>
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
-</head>
-<body>
-    <div class="container mt-4">
-        <h1 class="text-center my-4">WELCOME To LYS Mart</h1>
-<!--        <div th:replace="header :: content"></div>-->
-        <div class="text-center md-4">
-            <button id="loadProductButton" class="btn btn-primary">상품 목록 조회</button>
-            <button id="addProductButton" class="btn btn-success">상품 추가</button>
-        </div>
-        <ul id="productsList" class="list-group">
-            <!-- 상품 목록 추가되는 곳 -->
-        </ul>
-        <script>
-            let isListVisible = false;
-            document.getElementById('loadProductButton').addEventListener('click', function (){
-                if(isListVisible){
-                    document.getElementById('productsList').style.display='none';
-                    isListVisible = false;
-                    return;
-                }
-
-                fetch('/api/products')
-                    .then(Response => {
-                        if(!Response.ok){
-                            throw new Error('Network response was not ok');
-                        }
-                        return Response.json();
-                    })
-                    .then(data => {
-                        const productsList = document.getElementById('productsList');
-                        productsList.innerHTML = '';
-
-                        data.products.forEach(product => {
-                            const listItem = document.createElement('li');
-                            listItem.className = 'list-group-item';
-                            listItem.innerHTML = `<span>${product.id}</span> - <span>${product.name}</span> - <span>${product.price}</span> - <span>${product.stock}</span>`;
-                            productsList.appendChild(listItem);
-                        });
-
-                        productsList.style.display = 'block';
-                        isListVisible = true;
-                    })
-                    .catch(error => {
-                        console.error('Error fetching products:', error);
-                    });
-            });
-        </script>
-    </div>
-</body>
-</html>
