+++ src/main/java/com/lys/shoppingmall/model/ProductRequest.java
