+++ src/main/resources/sql/init_products.sql
