+++ src/main/resources/templates/add-product.html
