+++ src/main/resources/templates/product-detail.html
