+++ src/main/resources/templates/products.html
