+++ src/main/resources/templates/update-product.html
