+++ src/test/java/com/lys/shoppingmall/JasyptConfigTest.java
